Crate srml_finality_tracker

SRML module that tracks the last finalized block, as perceived by block authors.

Structs

GenesisConfig

InherentDataProvider
Provider for inherent data.

Module

ReportLatency
The delay after which point things become suspicious.

WindowSize
The number of recent samples to keep from this chain. Default is n-100

Enums

Call

Constants

INHERENT_IDENTIFIER
The identifier for the finalnum inherent.

Traits

FinalizedInherentData
Auxiliary trait to extract finalized inherent data.

OnFinalizationStalled
Called when finalization stalled at a given number.

Trait