Trait srml_finality_tracker::OnFinalizationStalled

pub trait OnFinalizationStalled<N> {
    fn on_stalled(further_wait: N);
}

Called when finalization stalled at a given number.

Required methods

fn on_stalled(further_wait: N)

The parameter here is how many more blocks to wait before applying changes triggered by finality stalling.

Loading content...